chore(legacy): drop stale comments in DesignContext restore-CSS effect
Removed a block of self-contradicting "train of thought" comments around
the on-mount CSS restore effect. The original text declared "we can't
re-apply (no data) — clear it" immediately followed by "Actually, store
the CSS separately so we can restore it" — which is exactly what the
code does. Replaced with one accurate one-liner. No behavior change.
